useWebPlayback custom hook
